Implement volume conservation param_t struct.
