+++ hoomd/md/CMakeLists.txt
@@ -56,7 +56,7 @@ set(_md_sources module-md.cc
                    Thermostat.cc
                    TwoStepNVTAlchemy.cc
                    WallData.cc
-		   VolumeConservationMeshForceCompute.cc
+		               VolumeConservationMeshForceCompute.cc
                    ZeroMomentumUpdater.cc
                    )
 
@@ -98,7 +98,7 @@ set(_md_headers ActiveForceComputeGPU.h
                 EvaluatorSpecialPairLJ.h
                 EvaluatorSpecialPairCoulomb.h
                 EvaluatorExternalElectricField.h
-		EvaluatorExternalMagneticField.h
+            		EvaluatorExternalMagneticField.h
                 EvaluatorExternalPeriodic.h
                 EvaluatorPairALJ.h
                 EvaluatorPairBuckingham.h
@@ -141,8 +141,8 @@ set(_md_headers ActiveForceComputeGPU.h
                 HarmonicDihedralForceCompute.h
                 HarmonicImproperForceComputeGPU.h
                 HarmonicImproperForceCompute.h
-		HelfrichMeshForceComputeGPU.h
-		HelfrichMeshForceCompute.h
+            		HelfrichMeshForceComputeGPU.h
+            		HelfrichMeshForceCompute.h
                 IntegrationMethodTwoStep.h
                 IntegratorTwoStep.h
                 ManifoldZCylinder.h
@@ -217,8 +217,9 @@ set(_md_headers ActiveForceComputeGPU.h
                 TwoStepConstantPressure.h
                 AlchemostatTwoStep.h
                 TwoStepNVTAlchemy.h
-		VolumeConservationMeshForceCompute.h
-		VolumeConservationMeshForceComputeGPU.h
+		            VolumeConservationMeshParameters.h
+		            VolumeConservationMeshForceCompute.h
+            		VolumeConservationMeshForceComputeGPU.h
                 AlchemostatTwoStep.h
                 WallData.h
                 ZeroMomentumUpdater.h
@@ -243,7 +244,7 @@ list(APPEND _md_sources ActiveForceComputeGPU.cc
                            HarmonicAngleForceComputeGPU.cc
                            HarmonicDihedralForceComputeGPU.cc
                            HarmonicImproperForceComputeGPU.cc
-			   HelfrichMeshForceComputeGPU.cc
+                  			   HelfrichMeshForceComputeGPU.cc
                            MolecularForceCompute.cu
                            NeighborListGPU.cc
                            NeighborListGPUBinned.cc
@@ -257,7 +258,7 @@ list(APPEND _md_sources ActiveForceComputeGPU.cc
                            TriangleAreaConservationMeshForceComputeGPU.cc
                            TwoStepBDGPU.cc
                            TwoStepLangevinGPU.cc
-			   VolumeConservationMeshForceComputeGPU.cc
+                  			   VolumeConservationMeshForceComputeGPU.cc
                            TwoStepConstantVolumeGPU.cc
                            TwoStepConstantPressureGPU.cc
                            MuellerPlatheFlowGPU.cc
@@ -271,7 +272,7 @@ set(_md_cu_sources ActiveForceComputeGPU.cu
                       AnisoPotentialPairALJ3GPUKernel.cu
                       AnisoPotentialPairDipoleGPUKernel.cu
                       AnisoPotentialPairGBGPUKernel.cu
-		      BendingRigidityMeshForceComputeGPU.cu
+            		      BendingRigidityMeshForceComputeGPU.cu
                       ComputeThermoGPU.cu
                       ComputeThermoHMAGPU.cu
                       ConstantForceComputeGPU.cu
@@ -283,7 +284,7 @@ set(_md_cu_sources ActiveForceComputeGPU.cu
                       HarmonicAngleForceGPU.cu
                       HarmonicDihedralForceGPU.cu
                       HarmonicImproperForceGPU.cu
-		      HelfrichMeshForceComputeGPU.cu
+            		      HelfrichMeshForceComputeGPU.cu
                       MolecularForceCompute.cu
                       NeighborListGPUBinned.cu
                       NeighborListGPU.cu
@@ -302,7 +303,7 @@ set(_md_cu_sources ActiveForceComputeGPU.cu
                       TwoStepConstantVolumeGPU.cu
                       TwoStepNVEGPU.cu
                       TwoStepRATTLENVEGPU.cu
-		      VolumeConservationMeshForceComputeGPU.cu
+            		      VolumeConservationMeshForceComputeGPU.cu
                       MuellerPlatheFlowGPU.cu
                       CosineSqAngleForceGPU.cu
                       )
